Move ProfileField from Model to Entity
